+++ src/core/PathUserTypeface.h
@@ -29,7 +29,7 @@ class PathUserTypeface final : public UserTypeface {
   static std::shared_ptr<UserTypeface> Make(uint32_t builderID, const std::string& fontFamily,
                                             const std::string& fontStyle,
                                             const FontMetrics& fontMetrics, const Rect& fontBounds,
-                                            float unitsPerEm,
+                                            int unitsPerEm,
                                             const VectorProviderType& glyphPathProviders);
   size_t glyphsCount() const override;
 
@@ -44,7 +44,7 @@ class PathUserTypeface final : public UserTypeface {
  private:
   explicit PathUserTypeface(uint32_t builderID, const std::string& fontFamily,
                             const std::string& fontStyle, const FontMetrics& fontMetrics,
-                            const Rect& fontBounds, float unitsPerEm,
+                            const Rect& fontBounds, int unitsPerEm,
                             const VectorProviderType& glyphPathProviders);
 
   VectorProviderType glyphPathProviders = {};
